Fold logfire adapter into importer package
