+++ plugins/packages/logfire-importer/README.md
@@ -15,6 +15,21 @@ The importer rebuilds span hierarchy and conservatively identifies model and too
 
 See the [Logfire import guide](https://docs.zenml.io/kitaru/guides/import-logfire-traces) for accepted formats, parameters, deduplication behavior, and fidelity limits.
 
+## Adapter
+
+The package also ships an adapter that imports Logfire traces of wrapped agent runs. The adapter uses the Logfire SDK already configured in your process and the Kitaru connection from your environment. Set `KITARU_AGENT_ID` to the agent imported sessions are created under and `LOGFIRE_TOKEN` to the write token the SDK records traces with.
+
+The trace fetch goes through the Logfire Query API, which authenticates with a read token, a separate credential from the SDK's write token. Create one under your Logfire project settings and set it as `LOGFIRE_READ_TOKEN`. Then wrap your agent entrypoint in a `LogfireAdapter` and run it through the adapter.
+
+```python
+from kitaru_logfire_importer import LogfireAdapter
+
+adapter = LogfireAdapter()
+result = adapter.run(my_agent, "Hello")
+```
+
+The adapter runs the function inside a Logfire trace, waits for Logfire to finish ingesting the trace, fetches it, and imports it as one Kitaru session. Use `run_async` for async functions. When the trace does not complete within the completeness timeout, the adapter creates a failed session carrying the trace id.
+
 ## Links
 
 - [Kitaru documentation](https://docs.zenml.io/kitaru)
