+++ plugins/packages/logfire-importer/pyproject.toml
@@ -22,7 +22,11 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 keywords = ["ai-agents", "kitaru", "logfire", "observability", "traces"]
-dependencies = ["kitaru>=0.22.0"]
+dependencies = [
+    "httpx>=0.27,<1",
+    "kitaru>=0.23.0",
+    "logfire>=4.35,<5",
+]
 
 [project.urls]
 Homepage = "https://kitaru.ai"
