--- plugins/packages/logfire/CHANGELOG.md
@@ -1,5 +1,0 @@
-# Changelog
-
-## 0.1.0
-
-- Initial release of the Logfire importer-backed adapter.
