--- plugins/packages/logfire/pyproject.toml
@@ -1,47 +1,0 @@
-[project]
-name = "kitaru-logfire"
-version = "0.1.0"
-description = "Logfire importer-backed adapter for Kitaru."
-readme = "README.md"
-license = "Apache-2.0"
-requires-python = ">=3.11"
-authors = [
-    { name = "ZenML GmbH", email = "info@zenml.io" },
-]
-classifiers = [
-    "Development Status :: 3 - Alpha",
-    "Intended Audience :: Developers",
-    "Intended Audience :: Science/Research",
-    "Operating System :: OS Independent",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.13",
-    "Programming Language :: Python :: 3.14",
-    "Topic :: Scientific/Engineering :: Artificial Intelligence",
-    "Topic :: Software Development :: Libraries :: Python Modules",
-]
-keywords = ["ai-agents", "kitaru", "logfire", "observability", "traces"]
-dependencies = [
-    "httpx>=0.27,<1",
-    "kitaru>=0.23.0",
-    "kitaru-logfire-importer>=0.1.1",
-    "logfire>=4.35,<5",
-]
-
-[project.urls]
-Homepage = "https://kitaru.ai"
-Documentation = "https://docs.zenml.io/kitaru/adapters/logfire"
-Repository = "https://github.com/zenml-io/kitaru"
-Issues = "https://github.com/zenml-io/kitaru/issues"
-Changelog = "https://github.com/zenml-io/kitaru/blob/develop/plugins/packages/logfire/CHANGELOG.md"
-
-[build-system]
-requires = ["hatchling==1.32.0"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/kitaru_logfire"]
-
-[tool.kitaru.artifact]
-import-module = "kitaru_logfire"
